Other files changed by move to all package (mainly reformatting tests)
