+++ src/test/java/all/useCase/login/LoginResponseModelUnitTests.java
@@ -1,4 +1,4 @@
-package useCase.login;
+package all.useCase.login;
 
 import org.junit.jupiter.api.Test;
 
@@ -8,12 +8,12 @@ public class LoginResponseModelUnitTests {
     void testLoginResponseModelShouldLogin() {
         LoginResponseModel responseModel = new LoginResponseModel(new LoginRequestModel("clare",
                 "12345"));
-        assert (responseModel.getUsername().equals("clare"));
-        assert (responseModel.getPassword().equals("12345"));
-        assert (responseModel.getId() == 1);
-        assert (responseModel.getEmail().equals("clare@gmail.com"));
-        assert (responseModel.getReviews().isEmpty());
-        assert (responseModel.getListings().isEmpty());
+        assert (responseModel.getUser().getUsername().equals("clare"));
+        assert (responseModel.getUser().getPassword().equals("12345"));
+        assert (responseModel.getUser().getID() == 1);
+        assert (responseModel.getUser().getEmail().equals("clare@gmail.com"));
+        assert (responseModel.getUser().getReviews().isEmpty());
+        assert (responseModel.getUser().getListings().isEmpty());
         //assert(responseModel.getCart().isEmpty());
     }
