+++ src/test/java/all/useCase/writeReview/ReviewCreatorUnitTests.java
@@ -1,9 +1,8 @@
-package writeReview;
+package all.useCase.writeReview;
 
-import entities.User;
-import features.Cart;
+import all.entities.User;
+import all.entities.Cart;
 import org.junit.jupiter.api.Test;
-import useCase.writeReview.ReviewCreator;
 
 import java.util.ArrayList;
