+++ src/test/java/all/useCase/writeReview/ReviewResponseModelUnitTests.java
@@ -1,8 +1,6 @@
-package writeReview;
+package all.useCase.writeReview;
 
 import org.junit.jupiter.api.Test;
-import useCase.writeReview.ReviewRequestModel;
-import useCase.writeReview.ReviewResponseModel;
 
 public class ReviewResponseModelUnitTests {
     ReviewRequestModel userExistsRequestModel = new ReviewRequestModel("user1",
